--update
-video backgroud
-import moviepy
-ALMOST ALL THE CODE FUCC

+++ WORDLE_SEM2/main.py
@@ -1,52 +1,51 @@
 import pygame
-import pygame.mixer
+import moviepy.editor as mp
 import os
 import sys
 import random
 from words import *
-#ini nat tes
-#ini jason tes 
+
 pygame.init()
 pygame.mixer.init()
-# ASSET DISPLAY AND SFX SEMUA
+
+# Constants
 WIDTH, HEIGHT = 650, 750
-SCREEN = pygame.display.set_mode((WIDTH, HEIGHT))
-BACKGROUND = pygame.image.load("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\tiles.png")
-#BACKGROUND = pygame.image.load("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\tiles.png")
-BACKGROUND_RECT = BACKGROUND.get_rect(center=(210, 290))
-ICON = pygame.image.load("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\Icon.png")
-# ICON = pygame.image.load("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\Icon.png")
-ENTER_SOUND = pygame.mixer.Sound("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\button_1.ogg")
-# ENTER_SOUND = pygame.mixer.Sound("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\button_1.ogg")
-TYPE_SOUND = pygame.mixer.Sound("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\type.ogg")
-# TYPE_SOUND = pygame.mixer.Sound("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\type.ogg")
-pygame.display.set_caption("projek apa ni")
-pygame.display.set_icon(ICON)
-BACKGROUND = pygame.image.load("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bali.jpg")
-# BACKGROUND = pygame.image.load("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bg_test.jpeg")
-BACKGROUND = pygame.transform.scale(BACKGROUND, (WIDTH, HEIGHT))
+LETTER_X_SPACING = 85
+LETTER_Y_SPACING = 12
+LETTER_SIZE = 75
 GREEN = "#6aaa64"
 YELLOW = "#c9b458"
 GREY = "#787c7e"
 OUTLINE = "#d3d6da"
 FILLED_OUTLINE = "#878a8c"
 CORRECT_WORD = "bread"
-# ALPHABET = ["ABCDEFGH", "IJKLMNOP", "QRSTUVWXYZ"]
 ALPHABET = ["ABCDEFGHIJKLM", "NOPQRSTUVWXYZ"]
+GUESSED_LETTER_FONT_PATH = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf"
+AVAILABLE_LETTER_FONT_PATH = GUESSED_LETTER_FONT_PATH
 
-GUESSED_LETTER_FONT = pygame.font.Font("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf", 50)
-AVAILABLE_LETTER_FONT = pygame.font.Font("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf", 25)
-# GUESSED_LETTER_FONT = pygame.font.Font("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf", 50)
-# AVAILABLE_LETTER_FONT = pygame.font.Font("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf", 25)
-SCREEN.blit(BACKGROUND, BACKGROUND_RECT)
-pygame.display.update()
-LETTER_X_SPACING = 85
-LETTER_Y_SPACING = 12
-LETTER_SIZE = 75
+# Setup display and video
+SCREEN = pygame.display.set_mode((WIDTH, HEIGHT))
+pygame.display.set_caption("projek apa ni")
 
-# VARIABLE GLOBAL
+video_path = "C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\bg_cth.mp4"
+video = mp.VideoFileClip(video_path)
+audio_path = "temp_audio.mp3"
+video.audio.write_audiofile(audio_path)
+pygame.mixer.music.load(audio_path)
+pygame.mixer.music.play(-1)
+
+ICON = pygame.image.load("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\Icon.png")
+pygame.display.set_icon(ICON)
+
+ENTER_SOUND = pygame.mixer.Sound("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\button_1.ogg")
+TYPE_SOUND = pygame.mixer.Sound("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\sfx\\type.ogg")
+
+GUESSED_LETTER_FONT = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 50)
+AVAILABLE_LETTER_FONT = pygame.font.Font(AVAILABLE_LETTER_FONT_PATH, 25)
+
+# Global variables
 guesses_count = 0
-guesses = [[]] * 5
+guesses = [[] for _ in range(5)]
 current_guess = []
 current_guess_string = ""
 current_letter_bg_x = 110
@@ -62,7 +61,7 @@ class Letter:
         self.bg_y = bg_position[1]
         self.bg_rect = (self.bg_x, self.bg_y + 120, LETTER_SIZE, LETTER_SIZE - 10)
         self.text = text
-        self.text_position = (self.bg_x + 36, self.bg_y + 150)
+        self.text_position = (self.bg_x + 37, self.bg_y + 150)
         self.text_surface = GUESSED_LETTER_FONT.render(self.text, True, self.text_color)
         self.text_rect = self.text_surface.get_rect(center=self.text_position)
 
@@ -72,12 +71,10 @@ class Letter:
             pygame.draw.rect(SCREEN, FILLED_OUTLINE, self.bg_rect, 3, border_radius=10)
         self.text_surface = GUESSED_LETTER_FONT.render(self.text, True, self.text_color)
         SCREEN.blit(self.text_surface, self.text_rect)
-        pygame.display.update()
 
     def delete(self):
         pygame.draw.rect(SCREEN, "white", self.bg_rect, border_radius=10)
         pygame.draw.rect(SCREEN, OUTLINE, self.bg_rect, 3, border_radius=10)
-        pygame.display.update()
 
 class Indicator:
     def __init__(self, x, y, letter):
@@ -88,38 +85,23 @@ class Indicator:
         self.bg_color = OUTLINE
 
     def draw(self):
-        pygame.draw.rect(SCREEN, self.bg_color, self.rect, border_radius=10)
+        pygame.draw.rect(SCREEN, self.bg_color, self.rect, border_radius=0)
         self.text_surface = AVAILABLE_LETTER_FONT.render(self.text, True, "black")
-        self.text_rect = self.text_surface.get_rect(center=(self.x + 27, self.y + 30))
+        self.text_rect = self.text_surface.get_rect(center=(self.x + 25, self.y + 25))
         SCREEN.blit(self.text_surface, self.text_rect)
-        pygame.display.update()
 
-
+# Create indicators
 indicator_x, indicator_y = 40, 0
-
 for row in ALPHABET:
     total_row_width = len(row) * LETTER_X_SPACING
-    indicator_x = 30 //(WIDTH - total_row_width)// 2 
+    indicator_x = -2
     for letter in row:
         new_indicator = Indicator(indicator_x, indicator_y, letter)
         indicators.append(new_indicator)
-        new_indicator.draw()
         indicator_x += 50
     indicator_y += 50
-# indicator_x, indicator_y = 20, 600
-
-# for row in ALPHABET:
-#     total_row_width = len(row) * LETTER_X_SPACING
-#     indicator_x = (WIDTH - total_row_width)// 2 
-#     for letter in row:
-#         new_indicator = Indicator(indicator_x, indicator_y, letter)
-#         indicators.append(new_indicator)
-#         new_indicator.draw()
-#         indicator_x += LETTER_X_SPACING
-#     indicator_y += 75
 
 def check_guess(guess_to_check):
-    # Goes through each letter and checks if it should be green, yellow, or grey.
     global current_guess, current_guess_string, guesses_count, current_letter_bg_x, game_result
     game_decided = False
     for i in range(5):
@@ -153,7 +135,6 @@ def check_guess(guess_to_check):
             game_result = ""
             game_decided = True
         guess_to_check[i].draw()
-        pygame.display.update()
     
     guesses_count += 1
     current_guess = []
@@ -164,39 +145,36 @@ def check_guess(guess_to_check):
         game_result = "L"
 
 def play_again():
-    # Puts the play again text on the screen.
     pygame.draw.rect(SCREEN, "grey", (10, 600, 1000, 600))
-    play_again_font = pygame.font.Font("D:\\Python_FInal_Project_Wordle\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf", 50)
-    # play_again_font = pygame.font.Font("C:\\Users\\Angeline\\Documents\\GitHub\\WORDLE_PROJ_SEM2\\WORDLE_SEM2\\assets\\FreeSansBold.otf", 50)
+    play_again_font = pygame.font.Font(GUESSED_LETTER_FONT_PATH, 50)
     play_again_text = play_again_font.render("Press ENTER to Play Again!", True, "black")
     play_again_rect = play_again_text.get_rect(center=(WIDTH/2, 700))
     word_was_text = play_again_font.render(f"The word was {CORRECT_WORD}!", True, "black")
     word_was_rect = word_was_text.get_rect(center=(WIDTH/2, 650))
     SCREEN.blit(word_was_text, word_was_rect)
     SCREEN.blit(play_again_text, play_again_rect)
-    pygame.display.update()
 
 def reset():
-    # Resets all global variables to their default states.
     global guesses_count, CORRECT_WORD, guesses, current_guess, current_guess_string, game_result
     SCREEN.fill("grey")
-    SCREEN.blit(BACKGROUND, BACKGROUND_RECT)
+    first_frame = video.get_frame(0)
+    first_frame_surface = pygame.surfarray.make_surface(first_frame.swapaxes(0, 1))
+    SCREEN.blit(first_frame_surface, (0, 0))
     guesses_count = 0
     CORRECT_WORD = random.choice(WORDS)
-    guesses = [[]] * 5
+    guesses = [[] for _ in range(5)]
     current_guess = []
     current_guess_string = ""
     game_result = ""
-    pygame.display.update()
     for indicator in indicators:
         indicator.bg_color = OUTLINE
         indicator.draw()
-
-def create_new_letter():
-    # Creates a new letter and adds it to the guess.
+    pygame.mixer.music.play(-1) 
+    
+def create_new_letter(key_pressed):
     global current_guess_string, current_letter_bg_x
     current_guess_string += key_pressed
-    new_letter = Letter(key_pressed, (current_letter_bg_x, guesses_count*75+LETTER_Y_SPACING))
+    new_letter = Letter(key_pressed, (current_letter_bg_x, guesses_count * 75 + LETTER_Y_SPACING))
     current_letter_bg_x += LETTER_X_SPACING
     guesses[guesses_count].append(new_letter)
     current_guess.append(new_letter)
@@ -205,17 +183,27 @@ def create_new_letter():
             letter.draw()
 
 def delete_letter():
-    # Deletes the last letter from the guess.
     global current_guess_string, current_letter_bg_x
-    guesses[guesses_count][-1].delete()
-    guesses[guesses_count].pop()
-    current_guess_string = current_guess_string[:-1]
-    current_guess.pop()
-    current_letter_bg_x -= LETTER_X_SPACING
+    if len(guesses[guesses_count]) > 0:
+        guesses[guesses_count][-1].delete()
+        guesses[guesses_count].pop()
+        current_guess_string = current_guess_string[:-1]
+        current_guess.pop()
+        current_letter_bg_x -= LETTER_X_SPACING
+
+clock = pygame.time.Clock()
+fps = video.fps
+start_time = pygame.time.get_ticks()
 
 while True:
-    if game_result != "":
-        play_again()
+    current_time = (pygame.time.get_ticks() - start_time) / 1000.0
+
+    if current_time < video.duration:
+        frame = video.get_frame(current_time)
+        frame_surface = pygame.surfarray.make_surface(frame.swapaxes(0, 1))
+        SCREEN.blit(frame_surface, (0, 0))
+    else:
+        start_time = pygame.time.get_ticks()
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
             pygame.quit()
@@ -228,12 +216,26 @@ while True:
                     if len(current_guess_string) == 5 and current_guess_string.lower() in WORDS:
                         check_guess(current_guess)
                         ENTER_SOUND.play()
+                        if game_result != "":
+                            play_again()
             elif event.key == pygame.K_BACKSPACE:
                 if len(current_guess_string) > 0:
                     delete_letter()
             else:
                 key_pressed = event.unicode.upper()
-                if key_pressed in "QWERTYUIOPASDFGHJKLZXCVBNM" and key_pressed != "":
+                if key_pressed in "QWERTYUIOPASDFGHJKLZXCVBNM":
                     if len(current_guess_string) < 5:
-                        create_new_letter()
+                        create_new_letter(key_pressed)
                         TYPE_SOUND.play()
+
+    for guess in guesses:
+        for letter in guess:
+            letter.draw()
+
+    for indicator in indicators:
+        indicator.draw()
+
+    pygame.display.update()
+    clock.tick(fps)
+
+os.remove(audio_path)
